+++ src/explorer/legacy/adapters/explorerAdapter.js
@@ -1,10 +1,10 @@
 // @flow
 
 import {type Node as ReactNode} from "react";
-import {Graph, type NodeAddressT} from "../../core/graph";
-import type {Assets} from "../../webutil/assets";
-import type {RepoId} from "../../core/repoId";
-import type {PluginDeclaration} from "../../analysis/pluginDeclaration";
+import {Graph, type NodeAddressT} from "../../../core/graph";
+import type {Assets} from "../../../webutil/assets";
+import type {RepoId} from "../../../core/repoId";
+import type {PluginDeclaration} from "../../../analysis/pluginDeclaration";
 
 export interface StaticExplorerAdapter {
   declaration(): PluginDeclaration;
