+++ src/explorer/legacy/pagerankTable/Connection.js
@@ -1,11 +1,11 @@
 // @flow
 
 import React from "react";
-import * as NullUtil from "../../util/null";
+import * as NullUtil from "../../../util/null";
 
-import type {NodeAddressT} from "../../core/graph";
-import type {Connection} from "../../core/attribution/graphToMarkovChain";
-import type {ScoredConnection} from "../../analysis/pagerankNodeDecomposition";
+import type {NodeAddressT} from "../../../core/graph";
+import type {Connection} from "../../../core/attribution/graphToMarkovChain";
+import type {ScoredConnection} from "../../../analysis/pagerankNodeDecomposition";
 import {DynamicExplorerAdapterSet} from "../adapters/explorerAdapterSet";
 import {TableRow} from "./TableRow";
 import {NodeRow} from "./Node";
